refactor(gda): lift the import-evidence adapter into a core import_evidence module (#741)
